Everything works with doubles

+++ MatrixCalc/src/main/java/matrixcalc/domain/Addition.java
@@ -12,17 +12,15 @@ public class Addition {
      * @param matrixB desc
      * @return the sum
      */
-    public static int[][] add(int[][] matrixA,int[][] matrixB ){
+    public static double[][] add(double[][] matrixA, double[][] matrixB ){
         
-        int[][] matrixC = new int[matrixA.length][matrixA.length];
+        double[][] matrixC = new double[matrixA.length][matrixA.length];
         
         for(int i = 0; i<matrixA.length; i++){
             for(int j = 0; j<matrixA.length; j++){
                matrixC[i][j] = matrixA[i][j] + matrixB[i][j]; 
             }
         }
-        
         return matrixC;
     }
-    
 }
